Enable accepting dialogs with Enter key (#68)
* Enable accepting dialogs with Enter key

Previously, the Quick Connect dialog did not allow users to press Enter
to confirm after typing a URI. This change sets a default widget for
several dialogs and allows certain entry rows to trigger the default
action when Enter is pressed.

This improves the usability and flow of the Quick Connect dialog.

* Add myself to developers